# perf: Speed up identifier splitting